Made game Ui elements into own class + fixed new implementation of GameUI in other classes + deleted duplicate of EventHandler
